version 23.2.1.0

+++ OpenDental/Forms/FormPaySimple.cs
@@ -98,7 +98,7 @@ namespace OpenDental {
 				textCheckSaveNumber.ReadOnly=true;
 				textBankName.ReadOnly=true;
 			}
-			if(_creditCard==null || !_creditCard.IsPaySimpleACH()) {
+			if(_creditCard==null || _creditCard.CCSource!=CreditCardSource.PaySimpleACH) {
 				textCardNumber.Select();
 				return;
 			}
@@ -109,7 +109,7 @@ namespace OpenDental {
 		private void FillFieldsFromCard() {
 			//User selected a credit card from drop down.
 			if(_creditCard.CCNumberMasked!="") {
-				if(_creditCard.IsPaySimpleACH()) {
+				if(_creditCard.CCSource==CreditCardSource.PaySimpleACH) {
 					textCheckSaveNumber.Text=_creditCard.CCNumberMasked;
 				}
 				else {
@@ -306,16 +306,69 @@ namespace OpenDental {
 				return null;
 			}
 			try {
-				apiResponseRetVal=ProcessPaymentHelper(expYear, expMonth);
-			}
-			catch(PaySimpleException ex) {
-				//Only handling PaySimpleException all other exceptions handled in ProcessPaymentHelperMethod
-				if(ex.ErrorType==PaySimpleError.CustomerDoesNotExist) {
-					PatientLinks.DeletePatNumTos(ex.CustomerId,PatientLinkType.PaySimple);
-					apiResponseRetVal=ProcessPaymentHelper(expYear, expMonth, true);
+				switch(_transType) {
+					case PaySimple.TransType.SALE:
+						//If _patCur is null or the PatNum is 0, we will make a one time payment for an UNKNOWN patient.  
+						//This is currently only intended for prepaid insurance cards.
+						apiResponseRetVal=PaySimple.MakePayment((_patient==null ? 0 : _patient.PatNum),_creditCard,PIn.Decimal(textAmount.Text),textCardNumber.Text
+							,new DateTime(expYear,expMonth,1),checkOneTimePayment.Checked,textZipCode.Text,textSecurityCode.Text,_clinicNum,_carrierName);
+						break;
+					case PaySimple.TransType.AUTH:
+						//Will retreive a new customer id from PaySimple if the patient doesn't exist already.
+						long paySimpleCustomerId=PaySimple.GetCustomerIdForPat(_patient.PatNum,_patient.FName,_patient.LName,_clinicNum);
+						//I have no idea if an insurance can make an auth payment but incase they can I check for it.
+						if(paySimpleCustomerId==0) {//Insurance payment, make a new customer id every time per Nathan on 04/26/2018
+							if((_patient==null || _patient.PatNum==0)) {
+								paySimpleCustomerId=PaySimple.AddCustomer("UNKNOWN","UNKNOWN","",_clinicNum);
+							}
+							else {
+								throw new ODException(Lan.g(this,"Invalid PaySimple Customer Id found."));
+							}
+						}
+						try {
+							apiResponseRetVal=PaySimple.AddCreditCard(paySimpleCustomerId,textCardNumber.Text,new DateTime(expYear,expMonth,1),textZipCode.Text,_clinicNum);
+						}
+						catch(PaySimpleException ex) {
+							PaySimple.HandlePaySimpleException(ex,paySimpleCustomerId);
+						}
+						break;
+					case PaySimple.TransType.RETURN:
+						if(string.IsNullOrWhiteSpace(textRefNumber.Text)) {
+							throw new ODException(Lan.g(this,"Invalid PaySimple Payment ID."));
+						}
+						if(!MsgBox.Show(this,MsgBoxButtons.OKCancel,"You are about to return a payment.  This action is irreversible.  Continue?")) {
+							throw new ODException(Lan.g(this,"Payment return was cancelled by user."));
+						}
+						apiResponseRetVal=PaySimple.ReversePayment(textRefNumber.Text,_clinicNum);
+						break;
+					case PaySimple.TransType.VOID:
+						if(string.IsNullOrWhiteSpace(textRefNumber.Text)) {
+							throw new ODException(Lan.g(this,"Invalid PaySimple Payment ID."));
+						}
+						if(!MsgBox.Show(this,MsgBoxButtons.OKCancel,"You are about to void a payment.  This action is irreversible.  Continue?")) {
+							throw new ODException(Lan.g(this,"Payment void was cancelled by user."));
+						}
+						apiResponseRetVal=PaySimple.VoidPayment(textRefNumber.Text,_clinicNum);
+						break;
+					default:
+						throw new Exception("Invalid transmission type: "+_transType.ToString());
 				}
 			}
-			if(apiResponseRetVal==null) { 
+			catch(PaySimpleException ex) {
+				MessageBox.Show(ex.Message);
+				if(ex.ErrorType==PaySimpleError.CustomerDoesNotExist && MsgBox.Show(this,MsgBoxButtons.OKCancel,
+					"Delete the link to the customer id for this patient?")) 
+				{
+					PatientLinks.DeletePatNumTos(ex.CustomerId,PatientLinkType.PaySimple);
+				}
+				return null;
+			}
+			catch(ODException wex) {
+				MessageBox.Show(wex.Message);//This should have already been Lans.g if applicable.
+				return null;
+			}
+			catch(Exception ex) {
+				MessageBox.Show(Lan.g(this,"Error:")+" "+ex.Message);
 				return null;
 			}
 			if(_transType.In(PaySimple.TransType.SALE,PaySimple.TransType.RETURN,PaySimple.TransType.VOID)) {//Only print a receipt if transaction is an approved SALE, RETURN, or VOID			
@@ -342,7 +395,7 @@ namespace OpenDental {
 			if(!checkOneTimePaymentACH.Checked
 				&& _creditCard!=null //if the user selected a saved CC
 				&& !string.IsNullOrWhiteSpace(_creditCard.PaySimpleToken) //there is a stored token for this card
-				&& _creditCard.IsPaySimpleACH())
+				&& _creditCard.CCSource==CreditCardSource.PaySimpleACH)
 			{
 				accountNumber=_creditCard.PaySimpleToken;
 			}
@@ -394,8 +447,8 @@ namespace OpenDental {
 			if(checkOneTimePaymentACH.Checked) {//not storing the account token
 				return apiResponseRetVal;
 			}
-			CreditCardSource source=_creditCard?.CCSource??CreditCardSource.PaySimpleACH;
-			UpsertCreditCard(apiResponseRetVal,StringTools.TruncateBeginning(textCheckSaveNumber.Text,4).PadLeft(textCheckSaveNumber.Text.Length,'*'),source,DateTime.MinValue);
+			UpsertCreditCard(apiResponseRetVal,StringTools.TruncateBeginning(textCheckSaveNumber.Text,4).PadLeft(textCheckSaveNumber.Text.Length,'*'),CreditCardSource.PaySimpleACH,
+				DateTime.MinValue);
 			return apiResponseRetVal;
 		}
 
@@ -404,7 +457,7 @@ namespace OpenDental {
 				_creditCard=new CreditCard();
 				_creditCard.IsNew=true;
 				_creditCard.PatNum=_patient.PatNum;
-				List<CreditCard> listCreditCardsItemOrderCount=CreditCards.RefreshAll(_patient.PatNum);
+				List<CreditCard> listCreditCardsItemOrderCount=CreditCards.Refresh(_patient.PatNum);
 				_creditCard.ItemOrder=listCreditCardsItemOrderCount.Count;
 			}
 			if(dateCCExp.Year > 1880) {
@@ -418,7 +471,6 @@ namespace OpenDental {
 				_creditCard.ClinicNum=_clinicNum;
 				_creditCard.Procedures=PrefC.GetString(PrefName.DefaultCCProcs);
 				CreditCards.Insert(_creditCard);
-				SecurityLogs.MakeLogEntry(EnumPermType.CreditCardEdit,_patient.PatNum,"Credit Card Added");
 			}
 			else {
 				CreditCards.Update(_creditCard);
@@ -486,7 +538,7 @@ namespace OpenDental {
 			if(textCardNumber.Text.Trim().Length<5) {
 				MsgBox.Show(this,"Invalid Card Number.");
 				return false;
-			}
+			}	
 			if(Regex.IsMatch(textExpDate.Text,@"^\d\d[/\- ]\d\d$")) {//08/07 or 08-07 or 08 07
 				try {//PIn.Int will throw an exception if not a valid format
 					expYear=PIn.Int("20"+textExpDate.Text.Substring(3,2));
@@ -530,10 +582,7 @@ namespace OpenDental {
 		}
 
 		private bool VerifyDataACH() {
-			// If Credit Card is not null, contains a token, and is currently set as a credit card of some sort in CCsource...
-			// AKA: If this thing is not an ACH account (but is/was a valid CreditCard)...
-			// Then validate the user input fields because apparently we are trying to convert the CC entry into an ACH.
-			if(string.IsNullOrEmpty(_creditCard?.PaySimpleToken) || !_creditCard.IsPaySimpleACH()) {
+			if(_creditCard==null || string.IsNullOrEmpty(_creditCard.PaySimpleToken) || _creditCard.CCSource!=CreditCardSource.PaySimpleACH) {
 				if(!Regex.IsMatch(textRoutingNumber.Text,"^[0-9]+$")) {
 					MsgBox.Show(this,"Invalid Routing Number.");
 					return false;
@@ -561,76 +610,7 @@ namespace OpenDental {
 			return true;
 		}
 
-		private PaySimple.ApiResponse ProcessPaymentHelper(int expYear,int expMonth,bool isRetry=false) {
-			PaySimple.ApiResponse apiResponseRetVal=null;
-			try {
-				switch(_transType) {
-					case PaySimple.TransType.SALE:
-						//If _patCur is null or the PatNum is 0, we will make a one time payment for an UNKNOWN patient.  
-						//This is currently only intended for prepaid insurance cards.
-						apiResponseRetVal=PaySimple.MakePayment((_patient==null ? 0 : _patient.PatNum),_creditCard,PIn.Decimal(textAmount.Text),textCardNumber.Text
-							,new DateTime(expYear,expMonth,1),checkOneTimePayment.Checked,textZipCode.Text,textSecurityCode.Text,_clinicNum,_carrierName);
-						break;
-					case PaySimple.TransType.AUTH:
-						//Will retreive a new customer id from PaySimple if the patient doesn't exist already.
-						long paySimpleCustomerId=PaySimple.GetCustomerIdForPat(_patient.PatNum,_patient.FName,_patient.LName,_clinicNum);
-						//I have no idea if an insurance can make an auth payment but incase they can I check for it.
-						if(paySimpleCustomerId==0) {//Insurance payment, make a new customer id every time per Nathan on 04/26/2018
-							if((_patient==null || _patient.PatNum==0)) {
-								paySimpleCustomerId=PaySimple.AddCustomer("UNKNOWN","UNKNOWN","",_clinicNum);
-							}
-							else {
-								throw new ODException(Lan.g(this,"Invalid PaySimple Customer Id found."));
-							}
-						}
-						try {
-							apiResponseRetVal=PaySimple.AddCreditCard(paySimpleCustomerId,textCardNumber.Text,new DateTime(expYear,expMonth,1),textZipCode.Text,_clinicNum);
-						}
-						catch(PaySimpleException ex) {
-							PaySimple.HandlePaySimpleException(ex,paySimpleCustomerId);
-						}
-						break;
-					case PaySimple.TransType.RETURN:
-						if(string.IsNullOrWhiteSpace(textRefNumber.Text)) {
-							throw new ODException(Lan.g(this,"Invalid PaySimple Payment ID."));
-						}
-						if(!MsgBox.Show(this,MsgBoxButtons.OKCancel,"You are about to return a payment.  This action is irreversible.  Continue?")) {
-							throw new ODException(Lan.g(this,"Payment return was cancelled by user."));
-						}
-						apiResponseRetVal=PaySimple.ReversePayment(textRefNumber.Text,_clinicNum);
-						break;
-					case PaySimple.TransType.VOID:
-						if(string.IsNullOrWhiteSpace(textRefNumber.Text)) {
-							throw new ODException(Lan.g(this,"Invalid PaySimple Payment ID."));
-						}
-						if(!MsgBox.Show(this,MsgBoxButtons.OKCancel,"You are about to void a payment.  This action is irreversible.  Continue?")) {
-							throw new ODException(Lan.g(this,"Payment void was cancelled by user."));
-						}
-						apiResponseRetVal=PaySimple.VoidPayment(textRefNumber.Text,_clinicNum);
-						break;
-					default:
-						throw new Exception("Invalid transmission type: "+_transType.ToString());
-				}
-			}
-			catch(PaySimpleException ex) {
-				if(ex.ErrorType==PaySimpleError.CustomerDoesNotExist && !isRetry) {
-					throw ex;
-				}
-				MessageBox.Show(ex.Message);
-				return null;
-			}
-			catch(ODException wex) {
-				MessageBox.Show(wex.Message);//This should have already been Lans.g if applicable.
-				return null;
-			}
-			catch(Exception ex) {
-				MessageBox.Show(Lan.g(this,"Error:")+" "+ex.Message);
-				return null;
-			}
-			return apiResponseRetVal;
-		}
-
-		private void butSave_Click(object sender,EventArgs e) {
+		private void butOK_Click(object sender,EventArgs e) {
 			Cursor=Cursors.WaitCursor;
 			if(tabControl.SelectedTab==tabCredit) {
 				int expYear;
@@ -657,6 +637,10 @@ namespace OpenDental {
 			else if(!_isAddingCard) {//If adding the card, leave the window open so the user can try again.
 				DialogResult=DialogResult.Cancel;
 			}
+		}
+
+		private void butCancel_Click(object sender,EventArgs e) {
+			DialogResult=DialogResult.Cancel;
 		}
 	}
 }
